Raise Add, Remove and Move SnippetTree events
SnippetContainers are implemented as linked lists and control all the insertion and removal logic.
They alone modify the next, previous and parentContainer properties of snippets. If a snippet is attached
to a SnippetTree the SnippetContainers are responsible to notify the SnippetTree. It is possible to append snippets
to other snippets before adding them to a snippet tree. This behaviour tries to mimic the DOM tree.

Snippet trees are in control of raising the Add, Remove and Move events. They make sure the snippetTree property of snippets
is set correctly.