Parse styles in Design
The styles defined in a livingdocs design are parsed in the Design class. For every template the styles defined globally, per group and per template are merged. The more specific styles overwrite the more general ones. E.g. a template style overwrites a group style with the same name. The styles are parsed and instantiated as DesignStyle instances. Identical DesignStyle instances are shared between templates.