Simplify SnippetModel#editables and #images
Separation of #editables and #images caused a lot of very similar code, so they are now merged in SnippetModel#content. This is part of a push to deal with directives in a more uniform way. #containers are special in a model and are still treated separately. This is in contrast to Template where all directives are stored in a DirectiveCollection. But I think this is justified due to the special role of SnippetContainers in SnippetModels.