Fix page#dragStart
Drags which started immediately (not delayed until a longpress) did not work anymore after pull-request #50. I register the move events immediately when someone calls dragStart, and this fixes the problem. Altough the coupling between page and DragDrop is too high at the moment and in need of a refactoring in my opinion.