Add #delayeUntilAttached to SnippetView
The image placeholder needs the dimensions of the image in the page.
So the placeholder method should only execute after the view is
inserted into the page.

The #delayUntilAttached will execute the given method once right after
the view is attached. There is also a cancelDelayed method to cancel
all pending methods for a directive so setting values multiple times
before a view is attached still works as expected.