Extend directives configuration
Convert directive configurations into literals with multiple properties
so it becomes easier to add new directives and to control their behaviour.
Unfortunately this made accessing the renderedAttribute for the directive
that will appear in the rendered html a lot longer.

Since this is used a lot in testing I had to change a lot of files for this change.
I introduced a test helper to turn something like
config.directives.editable.renderedAttr into test.editableAttr.
This is necessary since otherwise a lot of tests would become barely readable.