Separate template compilation and view creation
Until now the compilation of a template used the same method as the creation of a new view did.
This is only possible if no information is removed from the template html. The new code for optional
directives changes that.

Now for every new view the directives of a template are cloned and then linked with the Dom nodes
from the html of the new view. This makes creating a new view lighter since most of the parsing
and normalization of attributes is now only done in the template compilation.