Fix repeated split bug
The bug occurs because the model is not updated a the merge occurs.
Then when the user tries to split again the split action tries to
set the content of the model but this is ignored because the model
thinks the content has not changed.

Updating the model in the merge action resolved the issue.