Introduce placeholders for editables
The default values for editables are now shown via a css :before
pseudo-element. The SnippetView adds a data-attribute to the editable
directive with the default value when the editable is empty.

The data-attribute is also added if the editable is not empty or has focus
to prevent display bugs in Firefox where the height of the editable is collapsed
if there is focus on the editable while it has no content. To prevent this a
zero-width non-breaking space is set as content of the :before element.