Improve empty editable recognition
Sometimes it is not recognized correctly if an editable is emtpy.
To improve this the editable controller does some additional checks
on the value it reads from inside the editable to check if it is empty.

This logic should probably be moved into editableJS. There could be
a method to retrieve the value form editableJS which then could perform
cleanup.