Encapsulate render node in Page
The page now has a property called renderNode which contains the node
which Renderer will use to render to. The renderNode can be specified
during initialization. If no node is specified, a node with the default
CSS class will searched for and used instead.

The Renderer now only takes a page and gets the renderNode from the Page
that is passed into the initializer. If the page does not specify a
valid renderNode at this point, an error is raised.