Move editables initialization to page
This basically moves a some methods from the renderer to the page,
relieving the renderer of concerns that might not be present in other
types of pages, such as pages that don't allow editing.

The page is notified of a snippet view insertion. When notified, it
initializes the editables of that snippet view.