Fix failing doc-optional test
In SnippetView on content updates the optionals are shown or hidden
depending on their status. This is not animated.

When focusing or blurring a SnippetView different methods are used to
show or hide the optionals which use animations from config.

Moving the animations to CSS is not necessarily a good idea since
slideDown and slideUp are not so easy to accomplish.

With the current implementation optionals do not slideDown if the update
comes from the SnippetTree (for example in a preview view). This is not
so nice but we need to make sure optionals are visible immediately on the first
rendering. Thus it is simpler not to animate changes coming in from the SnippetTree
at all for the moment.