change data interface and add dataChanged event
The interface to the data is changed to set data when an object is
passed and get data when a string is passed. Also the method to
compare if the data is equal to a previous value is changed to a
very primitive deep equal.
The api exposes an event snippetDataChanged which is called with
the snippet and a list of all first-level data properties that
changed.