Fix bug #75
The method #html() is no longer responsible for calling render().

The method #render() does not empty the RenderingContainer anymore.
This should be in the responsibility of the RenderingContainer.
(Or be done once in the constructor)

There is a new isReady method so it becomes possible to check
if the renderer is ready (and thus already rendered).

I did not add a callback param to document#toHtml() to keep it simple.
The RenderingContainer should be ready immediately since it is just a
documentFragment. If that should not be true for some reason we throw
an error so we know at least what is going on.