Improve renderingContainer beforeReady
Always load a page asynchronously. This way it becomes less fragile
for subclasses to use the readySemaphore. With synchronous loading the
semaphore has to be incremented before it is started. And since this happens in the renderingContainer constructor it would have to happen before the call to
super(). But at this time the objects are not properly initialized.