improve metadata extraction on tree change
Until now, the entire tree was re-extracted when a component was
added, moved or removed and all fields were passed the the event
callbacks. Now, only the fields that could have changed are re-
extracted and only the fields that actually changed are passed to
the callback.

The callback is not called, if fields are not changed.